+++ lib/spout/models/option.rb
@@ -1,10 +1,10 @@
 module Spout
   module Models
     class Option
-      attr_accessor :display_name, :value, :description
+      attr_accessor :display_name, :value, :description, :missing
 
       def initialize(option_hash)
-        %w( display_name value description ).each do |method|
+        %w( display_name value description missing ).each do |method|
           instance_variable_set("@#{method}", (option_hash.kind_of?(Hash) ? option_hash : {})[method])
         end
       end