+++ lib/spout/models/table.rb
@@ -1,0 +1,14 @@
+
+
+
+module Spout
+  module Models
+    class Table
+
+
+      def initialize
+
+      end
+    end
+  end
+end